refactor: Move SymDebug to elf.rs
It's ELF-specific. Issue #1538
